+++ src/fabric/_types_powerbi.ts
@@ -1,0 +1,8 @@
+
+interface iPowerBiSqlEndpointSyncBatch {
+	batchId: string;
+	batchType: string;
+	endTimeStamp: Date;
+	progressState: string;
+	startTimeStamp: Date;
+}