+++ src/vscode/treeviews/Workspaces/FabricWorkspaceGenericViewer.ts
@@ -54,6 +54,14 @@ export class FabricWorkspaceGenericViewer extends FabricWorkspaceTreeItem {
 		}
 	}
 
+	get tempFilePath(): string {
+		let tempPath = this.apiPath.replace(/[^A-Za-z0-9\/:\.-]/g, "_");
+		if(this.apiPath.startsWith("https://")) {
+			tempPath = tempPath.replace("https://", "");
+		}
+		return tempPath;
+	}
+
 	public async showDefinition(): Promise<void> {
 		let result = await FabricApiService.get(this.apiPath);
 
@@ -66,7 +74,7 @@ export class FabricWorkspaceGenericViewer extends FabricWorkspaceTreeItem {
 			content = JSON.stringify(result.error, null, "\t")
 		}
 		
-		let tempUri = await TempFileSystemProvider.createTempFile(this.apiPath, content);
+		let tempUri = await TempFileSystemProvider.createTempFile(this.tempFilePath, content);
 
 		vscode.workspace.openTextDocument(tempUri).then(
 			document => vscode.window.showTextDocument(document)